requirements.txt
beautifulsoup4
chromadb
faiss
google-generativeai
langchain-text-splitters
lxml
numpy
pandas
Pillow
playwright
PyPDF2
python-docx
python-dotenv
rank-bm25
sentence-transformers
streamlit
torch
tqdm